bootstrap callcenter domains with ctable mapping
